moisture sensor now can be read

+++ src/sht1x.h
@@ -36,9 +36,9 @@ sht1x_error_t sht1x_status_read(uint8_t * data);
 
 sht1x_error_t sht1x_status_write(uint8_t status);
 
-sht1x_error_t sht1x_temperature_read(uint16_t * temperature);
+sht1x_error_t sht1x_temperature_read(float * temperature);
 
-sht1x_error_t sht1x_moisture_read(uint16_t * moisture);
+sht1x_error_t sht1x_moisture_read(float * moisture);
 
 sht1x_error_t sht1x_software_reset(void);
